# Shekinah Sons Movement Website  Welcome to the Shekinah Sons Movement website! This repository contains the code for our online presence.  ## About Us  **WE SHARE THE WORD OF GOD, WE ARE COMMITTED TO GROW IN THE KNOWLEDGE OF CHRIST.**
# Shekinah Sons Movement Website

Welcome to the Shekinah Sons Movement website! This repository contains the code for our online presence.

## About Us

**WE SHARE THE WORD OF GOD, WE ARE COMMITTED TO GROW IN THE KNOWLEDGE OF CHRIST.**

The Shekinah Sons Movement is a dynamic ministry dedicated to spreading the transformative power of the Gospel and fostering deep spiritual growth. We believe in the unwavering truth of God's Word and are passionate about equipping believers to live a life fully devoted to Jesus Christ.

Our ministry is built upon the foundation of love, grace, and the empowering presence of the Holy Spirit. We strive to create a welcoming and supportive community where individuals can experience the fullness of God's love and discover their purpose in Christ.

Through our various ministries, we aim to:

* **Teach and Equip:** Provide sound biblical teaching that empowers believers to grow in their faith and understanding of God's Word.
* **Worship and Praise:** Create an atmosphere of worship that honors God and invites His presence.
* **Reach Communities:** Engage with our local and global communities through outreach programs and service initiatives.
* **Nurture Spiritual Growth:** Offer resources and opportunities for individuals to deepen their relationship with God.

## Website Content

This website includes information about:

* Our teachings 
* Our ministries and activities (Agape Manifest, Awakened Dancers, Awakened Nation, etc.)
* Our music and media
* How to connect with us
* Technical information regarding our equipment and setup.
* Details regarding our ministrations.

## How to Use

To view the website, visit [rachael-mso.github.io/shekinah-sons-movement](rachael-mso.github.io/shekinah-sons-movement).




## Contact

For questions or inquiries, please contact us at:

* Email: shekinahsons@gmail.com
* Phone: +254 7022111560


Signed-off-by: Rachael-mso <133576495+Rachael-mso@users.noreply.github.com>